refactor: (BREAKING) Remove hive database and hive dependencies
