+++ test/fake_database.dart
@@ -16,33 +16,12 @@
  *   along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import 'package:file/local.dart';
-import 'package:hive/hive.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
 import 'package:matrix/matrix.dart';
 
 Future<DatabaseApi> getDatabase(Client? c, {String? databasePath}) =>
     getMatrixSdkDatabase(c, path: databasePath);
-
-bool hiveInitialized = false;
-
-// ignore: deprecated_member_use_from_same_package
-Future<HiveCollectionsDatabase> getHiveCollectionsDatabase(Client? c) async {
-  final testHivePath = await LocalFileSystem()
-      .systemTempDirectory
-      .createTemp('dart-sdk-tests-database');
-  if (!hiveInitialized) {
-    Hive.init(testHivePath.path);
-  }
-  // ignore: deprecated_member_use_from_same_package
-  final db = HiveCollectionsDatabase(
-    'unit_test.${c?.hashCode}',
-    testHivePath.path,
-  );
-  await db.open();
-  return db;
-}
 
 // ignore: deprecated_member_use_from_same_package
 Future<MatrixSdkDatabase> getMatrixSdkDatabase(
