+++ test_driver/matrixsdk_test.dart
@@ -18,7 +18,6 @@
 
 import 'dart:io';
 
-import 'package:hive/hive.dart';
 import 'package:olm/olm.dart' as olm;
 import 'package:test/test.dart';
 
@@ -40,8 +39,6 @@ void main() => group(
           Client? testClientA, testClientB;
 
           try {
-            Hive.init(null);
-
             await olm.init();
             olm.Account();
             Logs().i('[LibOlm] Enabled');
@@ -488,8 +485,6 @@ void main() => group(
           Client? testClientA, testClientB;
 
           try {
-            Hive.init(null);
-
             await olm.init();
             olm.Account();
             Logs().i('[LibOlm] Enabled');
